Change the execution order of step operations
Change the CPU so that every step first moves the instruction in the
target stage and then executes the given stage.

This is more compliant to what H&P state in Appendix C of CA:AQA, and
fixes issue #48.

It is a deep and risky change though, and there are a few hacks in there
(like the handling of RAW/WAW exceptions).

One thing that is broken for sure is the predict-not-taken behavior of
branches: branch instructions jump directly to the target, without
having the CPU first fetch the wrong instruction in case of taken
branch.

So for now I'll push to master the solution for #48 at commit
d3219c66361b3e1c19ec82b94e521222e0251667, and save this work in a
branch.

Maybe in the future I'll have time to test it more thoroughly.